Move Comparable and ColumnComparator to comparison package
