+++ fhirpath/src/main/java/au/csiro/pathling/fhirpath/collection/CodingCollection.java
@@ -31,8 +31,8 @@ import au.csiro.pathling.fhirpath.definition.NodeDefinition;
 import au.csiro.pathling.fhirpath.definition.def.DefCompositeDefinition;
 import au.csiro.pathling.fhirpath.definition.def.DefPrimitiveDefinition;
 import au.csiro.pathling.fhirpath.literal.CodingLiteral;
-import au.csiro.pathling.fhirpath.operator.ColumnComparator;
-import au.csiro.pathling.fhirpath.operator.Comparable;
+import au.csiro.pathling.fhirpath.comparison.ColumnComparator;
+import au.csiro.pathling.fhirpath.comparison.Comparable;
 import au.csiro.pathling.sql.misc.CodingToLiteral;
 import jakarta.annotation.Nonnull;
 import java.util.List;
