+++ fhirpath/src/main/java/au/csiro/pathling/fhirpath/collection/DateTimeCollection.java
@@ -25,8 +25,8 @@ import au.csiro.pathling.fhirpath.StringCoercible;
 import au.csiro.pathling.fhirpath.column.ColumnRepresentation;
 import au.csiro.pathling.fhirpath.column.DefaultRepresentation;
 import au.csiro.pathling.fhirpath.definition.NodeDefinition;
-import au.csiro.pathling.fhirpath.operator.ColumnComparator;
-import au.csiro.pathling.fhirpath.operator.Comparable;
+import au.csiro.pathling.fhirpath.comparison.ColumnComparator;
+import au.csiro.pathling.fhirpath.comparison.Comparable;
 import au.csiro.pathling.fhirpath.operator.DateTimeComparator;
 import au.csiro.pathling.fhirpath.operator.DefaultComparator;
 import au.csiro.pathling.sql.SqlFunctions;
