+++ fhirpath/src/main/java/au/csiro/pathling/fhirpath/comparison/CodingComparator.java
@@ -21,7 +21,6 @@ import static org.apache.spark.sql.functions.lit;
 import static org.apache.spark.sql.functions.when;
 
 import au.csiro.pathling.errors.InvalidUserInputError;
-import au.csiro.pathling.fhirpath.operator.ColumnComparator;
 import jakarta.annotation.Nonnull;
 import java.util.Arrays;
 import java.util.List;
